+++ src/core/marble/PlaybackWaitItem.cpp
@@ -8,21 +8,20 @@
 // Copyright 2014 Sanjiban Bairagya <sanjiban22393@gmail.com>
 //
 
-#include "PlaybackWaitItem.h"
-
-#include "GeoDataWait.h"
-
 #include <QTimer>
+
+#include "PlaybackWaitItem.h"
+#include "GeoDataWait.h"
 
 namespace Marble
 {
 
-PlaybackWaitItem::PlaybackWaitItem( const GeoDataWait* wait )
+PlaybackWaitItem::PlaybackWaitItem(const GeoDataWait *wait)
 {
     m_wait = wait;
     m_isPlaying = false;
 }
-const GeoDataWait* PlaybackWaitItem::wait() const
+const GeoDataWait *PlaybackWaitItem::wait() const
 {
     return m_wait;
 }
@@ -33,15 +32,21 @@ double PlaybackWaitItem::duration() const
 
 void PlaybackWaitItem::play()
 {
-    if( m_isPlaying ){
+    if (m_isPlaying)
+    {
         return;
-    } else {
+    }
+    else
+    {
         m_isPlaying = true;
-        if ( !( m_start.isValid() ) ){
+        if (!(m_start.isValid()))
+        {
             m_start = QDateTime::currentDateTime();
-            Q_ASSERT( m_start.isValid() );
-        } else {
-            m_start = m_start.addMSecs( m_pause.msecsTo( QDateTime::currentDateTime() ) );
+            Q_ASSERT(m_start.isValid());
+        }
+        else
+        {
+            m_start = m_start.addMSecs(m_pause.msecsTo(QDateTime::currentDateTime()));
         }
         playNext();
     }
@@ -49,18 +54,23 @@ void PlaybackWaitItem::play()
 
 void PlaybackWaitItem::playNext()
 {
-    if( !m_start.isValid() ){
+    if (!m_start.isValid())
+    {
         return;
     }
-    double const progress = m_start.msecsTo( QDateTime::currentDateTime() ) / 1000.0;
-    Q_ASSERT( progress >= 0.0 );
+    double const progress = m_start.msecsTo(QDateTime::currentDateTime()) / 1000.0;
+    Q_ASSERT(progress >= 0.0);
     double const t = progress / m_wait->duration();
-    if( t <= 1 ){
-        if( m_isPlaying ){
-            emit progressChanged( progress );
-            QTimer::singleShot( 20, this, SLOT(playNext()) );
+    if (t <= 1)
+    {
+        if (m_isPlaying)
+        {
+            emit progressChanged(progress);
+            QTimer::singleShot(20, this, SLOT(playNext()));
         }
-    } else {
+    }
+    else
+    {
         stop();
         emit finished();
     }
@@ -72,9 +82,9 @@ void PlaybackWaitItem::pause()
     m_pause = QDateTime::currentDateTime();
 }
 
-void PlaybackWaitItem::seek( double t )
+void PlaybackWaitItem::seek(double t)
 {
-    m_start = QDateTime::currentDateTime().addMSecs( -t * m_wait->duration() * 1000 );
+    m_start = QDateTime::currentDateTime().addMSecs(int(-t * m_wait->duration() * 1000));
     m_pause = QDateTime::currentDateTime();
 }
