+++ src/core/marble/SoundTrack.cpp
@@ -15,20 +15,20 @@
 namespace Marble
 {
 
-SoundTrack::SoundTrack( PlaybackSoundCueItem* item )
+SoundTrack::SoundTrack(PlaybackSoundCueItem *item)
 {
     m_item = item;
     m_progress = 0;
     m_delayBeforeTrackStarts = 0;
     m_paused = true;
-    connect( &m_timer, SIGNAL(timeout()), this, SLOT(playSlot()) );
+    connect(&m_timer, SIGNAL(timeout()), this, SLOT(playSlot()));
 }
 
-void SoundTrack::setDelayBeforeTrackStarts( double delay )
+void SoundTrack::setDelayBeforeTrackStarts(double delay)
 {
     m_delayBeforeTrackStarts = delay;
-    m_timer.setSingleShot( true );
-    m_timer.setInterval( m_delayBeforeTrackStarts * 1000 );
+    m_timer.setSingleShot(true);
+    m_timer.setInterval(int(m_delayBeforeTrackStarts * 1000));
 }
 
 double SoundTrack::delayBeforeTrackStarts() const
@@ -40,9 +40,12 @@ void SoundTrack::play()
 {
     m_paused = false;
     m_playTime = QDateTime::currentDateTime();
-    if( m_progress <= m_delayBeforeTrackStarts ){
-        m_timer.start( ( m_delayBeforeTrackStarts - m_progress ) * 1000 );
-    } else {
+    if (m_progress <= m_delayBeforeTrackStarts)
+    {
+        m_timer.start(int((m_delayBeforeTrackStarts - m_progress) * 1000));
+    }
+    else
+    {
         m_item->play();
     }
 }
@@ -56,35 +59,47 @@ void SoundTrack::pause()
 {
     m_paused = true;
     m_pauseTime = QDateTime::currentDateTime();
-    m_progress += m_playTime.msecsTo( m_pauseTime );
-    if( m_timer.isActive() ){
+    m_progress += m_playTime.msecsTo(m_pauseTime);
+    if (m_timer.isActive())
+    {
         m_timer.stop();
-    } else {
+    }
+    else
+    {
         m_item->pause();
     }
 }
 
-void SoundTrack::seek( double offset )
+void SoundTrack::seek(double offset)
 {
     m_timer.stop();
     m_progress = offset;
-    m_playTime = QDateTime::currentDateTime().addMSecs( -offset * 1000 );
+    m_playTime = QDateTime::currentDateTime().addMSecs(int(-offset * 1000));
 
-    if( offset <= m_delayBeforeTrackStarts ){
-        if( !m_paused ){
+    if (offset <= m_delayBeforeTrackStarts)
+    {
+        if (!m_paused)
+        {
             m_pauseTime = QDateTime();
             m_item->stop();
-            m_timer.start( ( m_delayBeforeTrackStarts - m_progress ) * 1000 );
-        } else {
+            m_timer.start(int((m_delayBeforeTrackStarts - m_progress) * 1000));
+        }
+        else
+        {
             m_pauseTime = QDateTime::currentDateTime();
         }
-    } else {
-        if( !m_paused ){
+    }
+    else
+    {
+        if (!m_paused)
+        {
             m_pauseTime = QDateTime();
-            m_item->seek( offset - m_delayBeforeTrackStarts );
-        } else {
+            m_item->seek(offset - m_delayBeforeTrackStarts);
+        }
+        else
+        {
             m_pauseTime = QDateTime::currentDateTime();
-            m_item->seek( offset - m_delayBeforeTrackStarts );
+            m_item->seek(offset - m_delayBeforeTrackStarts);
         }
     }
 }
@@ -94,12 +109,12 @@ void SoundTrack::stop()
     m_paused = true;
     m_item->stop();
     m_timer.stop();
-    m_playTime = QDateTime();
+    m_playTime  = QDateTime();
     m_pauseTime = QDateTime();
-    m_progress = 0;
+    m_progress  = 0;
 }
 
-void SoundTrack::setPaused( bool pause )
+void SoundTrack::setPaused(bool pause)
 {
     m_paused = pause;
 }
